requirements.txt
airflow
pillow